Added Readme and some info about the scripts

+++ uninstall-agent360.sh
@@ -46,9 +46,22 @@ if $symlink_hello ;then
 fi
 }
 
+check_ubuntu_release(){
+  OS_VERSION=$(cat /etc/os-release  | grep ^VERSION | head -1 | cut -d'"' -f2 | cut -d '.' -f1)
+  if [[ $OS_VERSION > 22 ]];then
+    echo -e "\\e[33m[WARNING] System version is ${OS_VERSION} applying arguments to pip3 --break-system-packages for package removal\\e[m"
+    handle_cmd 'systemctl stop agent360' 'The service agent360 has been stopped'
+    handle_cmd 'systemctl disable agent360' 'The service agent360 has been disabled'
+    handle_cmd 'pip3 uninstall -y --break-system-packages agent360' 'The Python modules for 360 Monitoring have been removed'
+    handle_cmd 'userdel agent360' 'The user agent360 has been deleted'
+  fi
+}
+## Added this function to check for the ubuntu version. Ubunu23+ requites --break-system-packages argument to install modules
+## Alternatively can be done via apt-get install python-agent360 but it is not yet added to any ubuntu repos
+check_ubuntu_release
 handle_cmd 'systemctl stop agent360' 'The service agent360 has been stopped'
 handle_cmd 'systemctl disable agent360' 'The service agent360 has been disabled'
-handle_cmd 'pip3 uninstall -y --break-system-packages agent360' 'The Python modules for 360 Monitoring have been removed'
+handle_cmd 'pip3 uninstall -y  agent360' 'The Python modules for 360 Monitoring have been removed'
 handle_cmd 'userdel agent360' 'The user agent360 has been deleted'
 
 if [[ -f /etc/systemd/system/agent360.service ]] || [[ -f /etc/systemd/system/agent360 ]] ; then
